+++ x/pocketcore/types/service.go
@@ -107,16 +107,16 @@ func (r *Relay) Validate(ctx sdk.Ctx, keeper PosKeeper, node nodeexported.Valida
 // "Execute" - Attempts to do a request on the non-native blockchain specified
 func (r Relay) Execute(hostedBlockchains *HostedBlockchains) (string, sdk.Error) {
 	// retrieve the hosted blockchain url requested
-	url, err := hostedBlockchains.GetChainURL(r.Proof.Blockchain)
+	chain, err := hostedBlockchains.GetChain(r.Proof.Blockchain)
 	if err != nil {
 		return "", err
 	}
-	url = strings.Trim(url, `/`)
+	url := strings.Trim(chain.URL, `/`)
 	if len(r.Payload.Path) > 0 {
 		url = url + "/" + strings.Trim(r.Payload.Path, `/`)
 	}
 	// do basic http request on the relay
-	res, er := executeHTTPRequest(r.Payload.Data, url, r.Payload.Method, r.Payload.Headers)
+	res, er := executeHTTPRequest(r.Payload.Data, url, chain.BasicAuth, r.Payload.Method, r.Payload.Headers)
 	if er != nil {
 		return res, NewHTTPExecutionError(ModuleName, er)
 	}
@@ -276,11 +276,14 @@ type DispatchResponse struct {
 }
 
 // "executeHTTPRequest" takes in the raw json string and forwards it to the RPC endpoint
-func executeHTTPRequest(payload string, url string, method string, headers map[string]string) (string, error) {
+func executeHTTPRequest(payload string, url string, basicAuth BasicAuth, method string, headers map[string]string) (string, error) {
 	// generate an http request
 	req, err := http.NewRequest(method, url, bytes.NewBuffer([]byte(payload)))
 	if err != nil {
-		return "", err
+		return "", errg
+	}
+	if basicAuth.Username != "" {
+		req.SetBasicAuth(basicAuth.Username, basicAuth.Password)
 	}
 	// add headers if needed
 	if len(headers) == 0 {